[minor] moved the badges to the center; change Camel case to snake_case
