+++ example/absorptive_memory/absorptive_experiment.py
