+++ example/absorptive_memory/absorptive_fidelity.py
