+++ example/absorptive_memory/absorptive_graph.py
