+++ example/absorptive_memory/absorptive_rate.py
