+++ example/absorptive_memory/absorptive_rate_graph.py
